+++ src/queryClauses/utils.ts
@@ -1,0 +1,7 @@
+import crypto from 'crypto'
+
+export const handleEmptyArray = (arr: unknown[]) => {
+	return arr.length === 0
+		? [crypto.randomUUID() + crypto.randomUUID() + crypto.randomUUID()]
+		: arr
+}