+++ src/types/cursor.ts
@@ -1,5 +1,5 @@
 import { MetaType } from './metaTypeCreator'
-import { ErrorMoreThanOnceDocSnapshotInCursor } from './error'
+import { ErrorMoreThanOnceDocSnapshotInCursor, ErrorEmptyCursor } from './error'
 import { IsUnion } from './utils'
 import { CursorConstraint } from './queryConstraints'
 import { DocumentSnapshot } from './snapshot'
@@ -7,7 +7,9 @@ import { DocumentSnapshot } from './snapshot'
 export type CursorType = 'startAt' | 'startAfter' | 'endAt' | 'endBefore'
 
 export type Cursor<Type extends CursorType> = <Values extends unknown[]>(
-	...snapshotOrFieldValues: Values extends (infer R)[]
+	...snapshotOrFieldValues: Values extends []
+		? ErrorEmptyCursor
+		: Values extends (infer R)[]
 		? DocumentSnapshot<MetaType> extends R
 			? IsUnion<R> extends false
 				? Values extends [infer Head, ...infer Rest]